Create weather entity

+++ pkg/temperature/converter.go
@@ -1,9 +1,9 @@
 package temperature
 
-func CelsiusToFahrenheit(c float32) float32 {
-	return (c*1.8) + 32
+func CelsiusToFahrenheit(c float64) float64 {
+	return (c * 1.8) + 32
 }
 
-func CelsiusToKelvin(c float32) float32 {
+func CelsiusToKelvin(c float64) float64 {
 	return c + 273
 }
